requirements.txt
beautifulsoup4==4.6.3
chardet==3.0.4
Django==2.1.4
django-background-tasks==1.2.0
django-compat==1.0.15
django-crispy-forms==1.7.2
djangorestframework==3.9.0
email-scrapper
lxml==4.2.5
pdfminer.six==20181108
pycryptodome==3.7.2
pytz==2018.7
six==1.12.0
sortedcontainers==2.1.0
